Adds getGeographyById(). Moves Geography to types.ts. Consolidates geography functions into geography.ts util file. Adds getParamStringArray(). Removes notion of ExtraParams from clipSketchToGeography() and instead uses geographyId. clipSketchToGeography() accepts simplification parameters.
